requirements.txt
pyairtable==2.3.3
jinja2==3.1.4
python-slugify==8.0.4
python-dotenv==1.0.1
markdown==3.6
pyyaml==6.0.1
